+++ src/widgets/TbButtonColumn.php
@@ -84,8 +84,8 @@ class TbButtonColumn extends CButtonColumn {
 			$options['title'] = $label;
 		}
 
-		if (!isset($options['data-toggle'])) {
-			$options['data-toggle'] = 'tooltip';
+		if (!isset($options['data-bs-toggle'])) {
+			$options['data-bs-toggle'] = 'tooltip';
 		}
 
 		if (isset($button['icon']) && $button['icon']) {