+++ src/widgets/TbToggleColumn.php
@@ -276,8 +276,8 @@ function() {
 		$htmlOptions = $this->button['htmlOptions'];
 
 		$htmlOptions['title'] = $label;
-		if (!isset($htmlOptions['data-toggle']))
-			$htmlOptions['data-toggle'] = 'tooltip';
+		if (!isset($htmlOptions['data-bs-toggle']))
+			$htmlOptions['data-bs-toggle'] = 'tooltip';
 
 		$iconHtml = TbIcon::render($icon);